Title: 입력받은 수의 합과 평균, Time: 149ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240508/입력받은 수의 합과 평균/sum-and-average-of-the-inputs.java
@@ -10,7 +10,7 @@ public class Main {
             int j=sc.nextInt();
             sum+=j;
         }
-        double avg=sum/n;
+        double avg=(double) sum/n;
         System.out.printf("%d %.1f",sum,avg);
     }
 }
